Add methods to Candidate model to deal with updates
The controller was starting to look a little fat, so I moved a couple of beefy update methods over to the model. This commit also includes the changes made to the controller to reflect those changes. The tests don't need changing because they only test the end result, and not the internals