Implement redaction
I've subclassed row, pulling out the method where personal data is added
to the Candidate object. I've replaced those attributions with a random
selection, where there has to be an object, or '[REDACTED]' where it
would be text. I've then broken out the collection of personal data to its own method so it can be overridden in the subclass RedactedRow. This allows for less code in the subclass and should therefore keep things slightly simpler in the future