Update profile page to show first and current role
I've made this change based on user feedback. As the initial dataset contains only the grade participants started at and their current roles, there was concern that some people might seem to have jumped significantly without any intervening roles. This change makes it clear that there was a gap!

I'm passing the 'first' and 'last' boolean variables on the loop to the template, so that it knows whether the role is the most recent or earliest. Kudos to Pallets for including that extremely helpful little counter, by the way. It means I can dynamically set which is the participants latest, versus earliest, role.