Use .apply method for dataframe
Iterating over rows in a dataframe is an anti-pattern. I've created a method in the Upload class, allowing me to use the .apply() method on the dataframe. This has sped up the tests by 30ms, which is not much in the grand scheme of things - but considering it's only reading one line it seems pretty impressive.