Add RoleChangeEvent table
When I started designing the API, I thought that a list of recent role changes would be an immensely valuable thing for another service to consume. Whether that was to generate a newsletter or identify trends in moves, knowing how frequently people were changing roles could be helpful. I then reflected on some of the incredibly complex code in this application and realised a lot of it centered around the Role table not being atomic - it had to carry data describing how the Candidate came to be in that role, which isn't necessarily useful and might not always be applicable. To that end I've created a new table and will now start to unpick the more complex code with a view to hopefully making the codebase simpler.